flags: move tests into package from main_test.go
